requirements.txt
streamlit>=1.31.0
google-generativeai
python-dotenv